libcuda-replay/requirements.txt
# SPDX-FileCopyrightText: 2021,2023 University of Rochester
#
# SPDX-License-Identifier: MIT

pycapnp
